+++ frontend/features/page/navigationMenu/NavigationBlock.tsx
