+++ frontend/features/page/navigationMenu/index.ts
