+++ frontend/features/page/sections/CardListSection/blocks/CardBlock.tsx
