+++ frontend/features/page/sections/FeatureListSection/blocks/FeatureBlock.tsx
