+++ frontend/features/page/sections/HeroSection/HeroSectionBlock.tsx
